+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+#### 3.0.0-beta5 - Unreleased
+* (Breaking Change) Added support for DateTimeOffset to JSON, XML, CSV, and HTML providers.
+* Map XSD date to System.DateTime and XSD dateTime to System.DateTimeOffset
+* Fixed large float values being silently converted to int.MinValue when parsing JSON.
+* Improved handling of invalid cookies.
+
 #### 3.0.0-beta4 - July 8 2018
 * (Breaking Change) Ignore culture when parsing JSON to better match the JSON spec.
 * Fixed handling of empty cookie headers.